+++ assets/js/check-updates.js
@@ -5,7 +5,7 @@
  * Receives configuration via wp_localize_script.
  *
  * @package SilverAssist\WpGithubUpdater
- * @version 1.3.1
+ * @version 1.4.0
  * @since 1.3.0
  */
 (($) => {